src/requirements.txt
google-genai
fastmcp
dotenv
mcpo
cloudinary
uv